build: split command catalog ui crate
