+++ crates/hermes-cli-ui/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "hermes-cli-ui"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Pure CLI presentation policy for Hermes Agent"
+
+[dependencies]
