+++ docs/parity/behavioral-similarity-diff.json
@@ -447,7 +447,7 @@
     "pass": true
   },
   "gate_failures": [],
-  "generated_at_utc": "2026-06-23T09:23:25.799787+00:00",
+  "generated_at_utc": "2026-06-23T10:14:45.420946+00:00",
   "manifest": "docs/parity/behavioral-similarity-cases.json",
   "schema_version": 1,
   "summary": {
