+++ docs/parity/deep-problem-solving-diff.json
@@ -440,7 +440,7 @@
     "pass": true
   },
   "gate_failures": [],
-  "generated_at_utc": "2026-06-23T09:23:28.203174+00:00",
+  "generated_at_utc": "2026-06-23T10:14:47.827059+00:00",
   "manifest": "docs/parity/deep-problem-solving-cases.json",
   "schema_version": 1,
   "summary": {
