+++ docs/parity/global-parity-proof.md
@@ -1,6 +1,6 @@
 # Global Parity Proof
 
-Generated: `2026-06-23T09:23:29.929691+00:00`
+Generated: `2026-06-23T10:14:49.130616+00:00`
 
 ## Gate Status
