+++ docs/parity/test-intent-mapping.md
@@ -1,6 +1,6 @@
 # Test Intent Mapping
 
-Generated: `2026-06-23T09:23:23.173757+00:00`
+Generated: `2026-06-23T10:14:42.605295+00:00`
 
 | Intent | Mapped | Evidence Count |
 | --- | --- | ---: |
