+++ docs/parity/workstream-status.json
@@ -1,6 +1,6 @@
 {
-  "generated_at_utc": "2026-06-23T02:03:48-06:00",
-  "head_sha": "af8ead255a7685b0e846df3eef6dd99f8e467ea2",
+  "generated_at_utc": "2026-06-23T03:25:28-06:00",
+  "head_sha": "b035fbba85aadb59051ed9172f65e6375e06d654",
   "states": {
     "complete": 7
   },
