+++ docs/parity/workstream-status.md
@@ -1,6 +1,6 @@
 # Workstream Status
 
-- Local HEAD: `af8ead255a7685b0e846df3eef6dd99f8e467ea2`
+- Local HEAD: `b035fbba85aadb59051ed9172f65e6375e06d654`
 - Upstream: `upstream/main` (`5ecf3bf0e0726b8b33682bb5c3aad9679b7b5be4`)
 
 | Workstream | Title | State |
