+++ scripts/audit-cargo-build-surface.sh
@@ -84,6 +84,13 @@ cargo tree -p hermes-cli --no-default-features --edges normal --depth 1
 echo '```'
 echo
 
+echo "## cargo tree: hermes-cli-ui --edges normal --depth 1"
+echo
+echo '```text'
+cargo tree -p hermes-cli-ui --edges normal --depth 1
+echo '```'
+echo
+
 echo "## cargo tree: hermes-parity-tests --edges dev --depth 1"
 echo
 echo '```text'
@@ -109,8 +116,9 @@ echo "## Interpretation"
 echo
 echo "- \`hermes-cli\` is currently the invalidation root for wrappers, the main runtime binary, TUI/clipboard UI dependencies, cron, ACP, MCP, tools, skills, and telemetry."
 echo "- Gateway adapter features are now behind explicit \`gateway-*\` CLI features. Default builds keep \`gateway-adapters-all\` for behavior compatibility, while \`--no-default-features\` checks compile the gateway core without adapter modules."
+echo "- Slash-command rendering, autocomplete ranking, alias canonicalization, and command catalog presentation now have a narrow dependency-free home in \`hermes-cli-ui\`."
 echo "- Provider/auth routing now has a narrower home in \`hermes-provider-runtime\`; agent configuration, query-mode provider/model/env/tool policy, model remediation, noninteractive agent-loop wiring, reply extraction, and runtime prompt reformulation policy now have a narrower home in \`hermes-app-runtime\`."
 echo "- Source/governance parity now lives in \`hermes-source-parity-tests\`, so command-contract checks avoid \`hermes-cli\`, \`clap\`, fixture-harness runtime crates, and protocol stack crates."
 echo "- Protocol differential parity now lives in \`hermes-protocol-parity-tests\`, isolating ACP/MCP/gateway/tool dependencies to tests that need them."
-echo "- The next high-value split is remaining reusable memory/context policy injection and tool-planning policy away from CLI/TUI, followed by extracting CLI UI/help presentation into a narrower module/crate."
+echo "- The next high-value split is remaining reusable memory/context policy injection and tool-planning policy away from CLI/TUI, followed by moving terminal UI and clipboard helpers into narrower presentation surfaces."
 echo "- Parity tests that only validate provider, auth, app-runtime, or command contracts should keep moving to narrower crates instead of pulling the full CLI binary surface."
